rename to respondent list info table
