+++ src/web/app/components/sessions-table/resend-results-link-to-student-modal/resend-results-link-to-student-modal.component.html
@@ -7,7 +7,7 @@
 </div>
 <div class="modal-body">
   <div class="table-responsive" *ngIf="studentListInfoTableRowModels.length > 0; else noStudentsToRemind">
-    <tm-student-list-info-table [(studentListInfoTableRowModels)]="studentListInfoTableRowModels" [shouldDisplayHasSubmittedSessionColumn]="false" [isStudent]="true"></tm-student-list-info-table>
+    <tm-respondent-list-info-table [(studentListInfoTableRowModels)]="studentListInfoTableRowModels" [shouldDisplayHasSubmittedSessionColumn]="false"></tm-respondent-list-info-table>
   </div>
   <ng-template #noStudentsToRemind>
     <h4 class="p-2 bg-info text-white">
@@ -15,7 +15,7 @@
     </h4>
   </ng-template>
   <div class="table-responsive" *ngIf="instructorListInfoTableRowModels.length > 0; else noInstructorsToRemind">
-    <tm-student-list-info-table [(instructorListInfoTableRowModels)]="instructorListInfoTableRowModels" [shouldDisplayHasSubmittedSessionColumn]="false" [isStudent]="false"></tm-student-list-info-table>
+    <tm-respondent-list-info-table [(instructorListInfoTableRowModels)]="instructorListInfoTableRowModels" [shouldDisplayHasSubmittedSessionColumn]="false"></tm-respondent-list-info-table>
   </div>
   <ng-template #noInstructorsToRemind>
     <h4 class="p-2 bg-info text-white">
