+++ src/web/app/components/sessions-table/resend-results-link-to-student-modal/resend-results-link-to-student-modal.component.spec.ts
@@ -5,7 +5,7 @@ import { FormsModule } from '@angular/forms';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
 
 import { MatSnackBarModule } from '@angular/material';
-import { StudentListInfoTableComponent } from '../student-list-info-table/student-list-info-table.component';
+import { RespondentListInfoTableComponent } from '../respondent-list-info-table/respondent-list-info-table.component';
 import { ResendResultsLinkToStudentModalComponent } from './resend-results-link-to-student-modal.component';
 
 @Component({ selector: 'tm-ajax-preload', template: '' })
@@ -20,7 +20,7 @@ describe('ResendResultsLinkToStudentModalComponent', () => {
       declarations: [
         ResendResultsLinkToStudentModalComponent,
         AjaxPreloadComponent,
-        StudentListInfoTableComponent,
+        RespondentListInfoTableComponent,
       ],
       imports: [
         HttpClientTestingModule,
