+++ src/web/app/components/sessions-table/respondent-list-info-table/respondent-list-info-table.component.scss
