+++ src/web/app/components/sessions-table/respondent-list-info-table/student-list-info-table-model.ts
