+++ src/web/app/components/sessions-table/send-reminders-to-student-modal/send-reminders-to-student-modal.component.spec.ts
@@ -5,7 +5,7 @@ import { Component } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { MatSnackBarModule } from '@angular/material';
 import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
-import { StudentListInfoTableComponent } from '../student-list-info-table/student-list-info-table.component';
+import { RespondentListInfoTableComponent } from '../respondent-list-info-table/respondent-list-info-table.component';
 import { SendRemindersToStudentModalComponent } from './send-reminders-to-student-modal.component';
 
 @Component({ selector: 'tm-ajax-preload', template: '' })
@@ -20,7 +20,7 @@ describe('SendRemindersToStudentModalComponent', () => {
       declarations: [
         SendRemindersToStudentModalComponent,
         AjaxPreloadComponent,
-        StudentListInfoTableComponent,
+        RespondentListInfoTableComponent,
       ],
       imports: [
         HttpClientTestingModule,
