+++ src/web/app/components/sessions-table/send-reminders-to-student-modal/send-reminders-to-student-modal.component.ts
@@ -3,7 +3,7 @@ import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';
 import {
   InstructorListInfoTableRowModel,
   StudentListInfoTableRowModel,
-} from '../student-list-info-table/student-list-info-table-model';
+} from '../respondent-list-info-table/student-list-info-table-model';
 
 /**
  * Send reminders to students modal.
